pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pixelprompt"
version = "0.1.0"
description = "Compress LLM context by rendering text as optimized images"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Gabriele Venturi", email = "gabriele@sinaptik.ai"},
]
keywords = ["llm", "context", "compression", "image", "token-optimization"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
]

dependencies = [
    "pillow>=10.0.0",
    "anthropic>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
]

[project.urls]
Homepage = "https://github.com/sinaptik-ai/pixelprompt"
Documentation = "https://pixelprompt.readthedocs.io"
Repository = "https://github.com/sinaptik-ai/pixelprompt"
Issues = "https://github.com/sinaptik-ai/pixelprompt/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/pixelprompt"]

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.ruff]
line-length = 100
target-version = "py38"
select = ["E", "F", "W", "I"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src/pixelprompt --cov-report=term-line"
